More work on an orthogonal browsing of stuff and renaming of things.
I'm trying to build a nice JSON model in a sort of REST like way that the user
can use to explore the data. For example rather than showing them all of the
possible options initially we let them drill down and pick stuff out and then
from there we can let them save common things, etc... But I also wanna keep
this logic in the data that's returned as JSON from the server.