Cleaning up how we read the databases in from disk.
DatabaseOnDisk now has a Category and the finder does the work of narrowing
those things down. Also use name instead of display_name.

Removing a bunch of unnecessary stuff as well.

Shit. I keep ammending when I use the shell to go back to a previous command
and now I've squashed a bunch of changes making thigns sorting and cleaning up
the model code by using a SortedSet instead of a plain array.