Got graph selection working.
Links show up as selected for the given graph thanks to CSS selectors and our
URIs per graph. I'm pretty happy with this, I wish things didn't flicker as
much when they changed though. Mostly this was just tweaking how the parameters
for the graph and the URIs were being manipulated.